+++ rush01/Assets/Scripts/Enemy/Enemy.cs
@@ -154,7 +154,7 @@ public class Enemy : MonoBehaviour {
 	}
 
 	public virtual void Damage () {
-		int val = 75 + agi - PlayerScript.instance.agi - Random.Range (1, 101);
+		float val = 75 + agi - PlayerScript.instance.agi - Random.Range (1, 101);
 		bool hit = val > 0 ? true : false;
 
 		if (intruder && Vector3.Distance (this.transform.position, intruder.transform.position) <= 2.0 && hit) {
